+++ clinicaOdontologica/src/main/java/com/digitalhouse/backend/integrador/clinicaOdontologica/entity/Odontologo.java
@@ -1,12 +1,12 @@
 package com.digitalhouse.backend.integrador.clinicaOdontologica.entity;
 
 public class Odontologo {
-    private int id;
+    private Long id;
     private int numeroMatricula;
     private String nombre;
     private String apellido;
 
-    public Odontologo(int id, int numeroMatricula, String nombre, String apellido) {
+    public Odontologo(Long id, int numeroMatricula, String nombre, String apellido) {
         this.id = id;
         this.numeroMatricula = numeroMatricula;
         this.nombre = nombre;
@@ -19,11 +19,11 @@ public class Odontologo {
         this.apellido = apellido;
     }
 
-    public int getId() {
+    public Long getId() {
         return this.id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
